+++ src/log.ts
@@ -1,17 +1,17 @@
 import * as vscode from "vscode";
 
-const output = vscode.window.createOutputChannel("Phabricator");
+const outputChannel = vscode.window.createOutputChannel("Phabricator");
 const append = (text: string) => {
   const now = new Date();
-  output.appendLine(
-    `[${[now.getHours(), now.getMinutes(), now.getSeconds()]
-      .map(value => (value + "").padStart(2, "0"))
-      .join(":")}] - ${text}`
-  );
+  const output = `[${[now.getHours(), now.getMinutes(), now.getSeconds()]
+    .map(value => (value + "").padStart(2, "0"))
+    .join(":")}] - ${text}`;
+  outputChannel.appendLine(output);
+  return output;
 };
 
 const show = () => {
-  output.show();
+  outputChannel.show();
 };
 
 export default {