+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 - [internal] Add EditorConfig file
+- [internal] Add CI tests
 
 ## [1.0.5] - 2020-01-18